split MolAlign lib into two pieces
